refactor(audio-bridge): results -> responses
