+++ jarust_plugins/src/audio_bridge/handle.rs
@@ -1,19 +1,34 @@
-use super::{
-    messages::{
-        AudioBridgeAction, AudioBridgeAllowedMsg, AudioBridgeAllowedOptions, AudioBridgeCreateMsg,
-        AudioBridgeCreateOptions, AudioBridgeDestroyMsg, AudioBridgeDestroyOptions,
-        AudioBridgeEditMsg, AudioBridgeEditOptions, AudioBridgeExistsMsg, AudioBridgeJoinMsg,
-        AudioBridgeJoinOptions, AudioBridgeKickAllMsg, AudioBridgeKickAllOptions,
-        AudioBridgeKickMsg, AudioBridgeKickOptions, AudioBridgeListMsg,
-        AudioBridgeListParticipantsMsg, AudioBridgeResumeMsg, AudioBridgeResumeOptions,
-        AudioBridgeSuspendMsg, AudioBridgeSuspendOptions,
-    },
-    results::{AudioBridgePluginData, AudioBridgePluginEvent, Participant, Room},
-};
+use super::messages::AudioBridgeAction;
+use super::messages::AudioBridgeAllowedMsg;
+use super::messages::AudioBridgeAllowedOptions;
+use super::messages::AudioBridgeCreateMsg;
+use super::messages::AudioBridgeCreateOptions;
+use super::messages::AudioBridgeDestroyMsg;
+use super::messages::AudioBridgeDestroyOptions;
+use super::messages::AudioBridgeEditMsg;
+use super::messages::AudioBridgeEditOptions;
+use super::messages::AudioBridgeExistsMsg;
+use super::messages::AudioBridgeJoinMsg;
+use super::messages::AudioBridgeJoinOptions;
+use super::messages::AudioBridgeKickAllMsg;
+use super::messages::AudioBridgeKickAllOptions;
+use super::messages::AudioBridgeKickMsg;
+use super::messages::AudioBridgeKickOptions;
+use super::messages::AudioBridgeListMsg;
+use super::messages::AudioBridgeListParticipantsMsg;
+use super::messages::AudioBridgeResumeMsg;
+use super::messages::AudioBridgeResumeOptions;
+use super::messages::AudioBridgeSuspendMsg;
+use super::messages::AudioBridgeSuspendOptions;
+use super::responses::AudioBridgePluginData;
+use super::responses::AudioBridgePluginEvent;
+use super::responses::Participant;
+use super::responses::Room;
 use jarust::japrotocol::EstablishmentProtocol;
 use jarust::jatask::AbortHandle;
 use jarust::prelude::*;
-use std::{ops::Deref, time::Duration};
+use std::ops::Deref;
+use std::time::Duration;
 
 pub struct AudioBridgeHandle {
     handle: JaHandle,
